refactor(template): remove redundant total talks display
Removed the conditional statement displaying the total number of talks when the filtered count is not equal to the total talks in the index.html template. This improves the clarity of the user interface by avoiding unnecessary information and streamlining the presentation of data related to user ratings and displayed talks.